requirements.txt
fastapi==0.95.1
twilio==8.1.0
langchain==0.0.150 
python-dotenv==1.0.0
openai==0.27.6
mangum==0.17.0
uvicorn==0.22.0
urllib3==1.26.15
python-multipart==0.0.6
boto3==1.26.129
pinecone-client==2.2.1
tiktoken==0.4.0